make Label offsets int
It's a bit too easy to trip up the MSRTC bots with size_t.

(The only reason I used size_t in the first place is that
these really are byte offsets... kind of the one place we
use size_t.)

Cq-Include-Trybots: skia.primary:Test-Win2016-MSVC-GCE-CPU-AVX2-x86_64-Debug-All-MSRTC

+++ src/core/SkVM.cpp
@@ -643,7 +643,7 @@ namespace skvm {
     void Assembler::vcvttps2dq(Ymm dst, Ymm x) { this->op(0xf3,0x0f,0x5b, dst,x); }
 
     Assembler::Label Assembler::here() {
-        return { this->size() };
+        return { (int)this->size() };
     }
 
     void Assembler::op(int prefix, int map, int opcode, Ymm dst, Ymm x, Label l) {
